% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_zip_demographics}
\alias{df_zip_demographics}
\title{A data.frame containing demographic statistics for each Zip Code Tabulated Area (ZCTA).}
\usage{
data(df_zip_demographics)
}
\description{
A data.frame containing demographic statistics for each Zip Code Tabulated Area (ZCTA).
}
\references{
Data comes from the 2013 5-year American Community Survey (ACS). Data generated by
?get_zip_demographics.
}
\keyword{data}

