% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_zip_demographics.R
\name{get_zip_demographics}
\alias{get_zip_demographics}
\title{Get a handful of demographic variables from the US Census Bureau as a data.frame.}
\usage{
get_zip_demographics(endyear = 2013, span = 5)
}
\arguments{
\item{endyear}{The end year for the survey}

\item{span}{The span of the survey}
}
\description{
The data comes from the American Community Survey (ACS). The geographic unit is
Zip Code Tabulated Areas (ZCTAs). The variables are: total population, percent White
not Hispanic, Percent Black or African American not Hispanic, percent Asian not Hispanic,
percent Hispanic all races, per-capita income, median rent and median age.
}
\references{
The choroplethr guide to Census data: http://cran.r-project.org/web/packages/choroplethr/vignettes/e-mapping-us-census-data.html

A list of all ACS Surveys: http://factfinder.census.gov/faces/affhelp/jsf/pages/metadata.xhtml?lang=en&type=survey&id=survey.en.ACS_ACS
}

