% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_choropleth}
\alias{zip_choropleth}
\title{Create a choropleth of USA Zip Code Tabulated Areas (ZCTAs) with sensible defaults.}
\usage{
zip_choropleth(df, title = "", legend = "", buckets = 7, zoom = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Elements in
the "region" column must exactly match how regions are named in the "region" column in zcta.map.}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.}

\item{zoom}{An optional vector of ZIPs to zoom in on. Elements of this vector must exactly
match the names of ZIPs as they appear in the "region" column of ?zcta.regions}
}
\description{
The map used is zcta.map.  See zcta.regions for a list of valid regions.
}

