% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_choropleth.R
\docType{data}
\name{ZipChoropleth}
\alias{ZipChoropleth}
\title{An R6 object for creating ZIP-level choropleths.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ZipChoropleth
}
\description{
An R6 object for creating ZIP-level choropleths.
}
\keyword{datasets}
