% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zip.map}
\alias{zip.map}
\title{Map of US Zip Code Tabulated Areas (ZCTAs)}
\usage{
data(zip.map)
}
\description{
The map
was modified using QGIS in order to remove Puerto Rico, islands off of Alaska that
crossed the antimeridian and American Samoa. A column named "region" was added for convenience which
is a duplicate of the column named ZCTA5CE10.
}
\references{
Taken from the US Census 2013 Cartographic Boundary Shapefiles - ZIP Code Tabulation Areas (ZCTAs):
https://www.census.gov/geo/maps-data/data/cbf/cbf_zcta.html.
}

