% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zip.regions}
\alias{zip.regions}
\title{A data.frame consisting of the name of each region (i.e. zip code) in the map zcta.map along with additional information.}
\usage{
data(zip.regions)
}
\description{
The additional information is the state name, county name, county fips code, cbsa code, cbsa title and
metropolitan/micropolitan statistical area name (if applicable).
}
\details{
Note that this data.frame contains duplicates, because some zip codes span counties and states.
}
\examples{
data(zip.regions)
head(zip.regions)
}
\references{
Data comes from the US Census Bureau's 2010 ZCTA to County Relationship File (
https://www.census.gov/geo/maps-data/data/zcta_rel_download.html), 2010 ZCTA to Metropolitan and Micropolitan Statistical
Areas Relationship File and the Core based statistical areas (CBSAs) and combined statistical areas (CSAs) Metropolitan
and Micropolitan Delineation files (http://www.census.gov/population/metro/data/def.html).
}
\seealso{
?zip.map
}

