% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zip.regions}
\alias{zip.regions}
\title{Metadata about US Zip Code Tabulated Areas (ZCTAs)}
\usage{
data(zip.regions)
}
\description{
The metadata is: State name, County name, County FIPS code, Core Based Statistical
Area (CBSA) id, CBSA title and 
Metropolitan/Micropolitan Statistical Area (MSA) name.
}
\details{
Note that this data.frame contains duplicates, because some ZCTAs span multiple Counties and States.
}
\references{
Data comes from the US Census Bureau's 2010 ZCTA to County Relationship File 
(https://www.census.gov/geo/maps-data/data/zcta_rel_download.html), 2010 ZCTA to Metropolitan and Micropolitan Statistical 
Areas Relationship File and the Core Based Statistical Areas (CBSAs) and Combined Statistical Areas (CSAs) Metropolitan 
and Micropolitan Delineation files (http://www.census.gov/population/metro/data/def.html).
}
\seealso{
?zip.map
}
