% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zip.regions}
\alias{zip.regions}
\title{A data.frame consisting of the name of each region in the map zcta.map.}
\usage{
data(zip.regions)
}
\description{
A data.frame consisting of the name of each region in the map zcta.map.
}
\examples{
data(zip.regions)
head(zip.regions)
}
\seealso{
?zip.map
}

