% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zip.R
\docType{data}
\name{ZipChoropleth}
\alias{ZipChoropleth}
\title{Create a ZIP-level choropleth}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x13b0c3958> 
 - attr(*, "name")= chr "ZipChoropleth_generator"
}}
\usage{
ZipChoropleth
}
\description{
Create a ZIP-level choropleth
}
\keyword{datasets}

